requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.0
bootstrap-flask==3.3.7.1
Flask-CKEditor==1.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
pandas==2.2.3
bs4==0.0.1
SQLAlchemy==2.0.21
Werkzeug==3.0.1
Flask_Sqlalchemy==3.1.1